create steps module
