+++ vscode/microsoft-kiota/src/modules/steps/index.ts
@@ -1,0 +1,215 @@
+import { Disposable, OpenDialogOptions, QuickInput, QuickInputButton, QuickInputButtons, QuickPickItem, Uri, window } from 'vscode';
+
+export interface BaseStepsState {
+    title: string;
+    step: number;
+    totalSteps: number;
+}
+
+class InputFlowAction {
+    static back = new InputFlowAction();
+    static cancel = new InputFlowAction();
+    static resume = new InputFlowAction();
+}
+
+type InputStep = (input: MultiStepInput) => Thenable<InputStep | void>;
+
+interface QuickPickParameters<T extends QuickPickItem> {
+    title: string;
+    step: number;
+    totalSteps: number;
+    items: T[];
+    activeItem?: T;
+    ignoreFocusOut?: boolean;
+    placeholder: string;
+    buttons?: QuickInputButton[];
+    shouldResume: () => Thenable<boolean>;
+}
+
+interface InputBoxParameters {
+    title: string;
+    step: number;
+    totalSteps: number;
+    value: string;
+    prompt: string;
+    validate?: (value: string) => Promise<string | undefined>;
+    buttons?: QuickInputButton[];
+    ignoreFocusOut?: boolean;
+    placeholder?: string;
+    shouldResume: () => Thenable<boolean>;
+}
+
+interface OpenDialogParameters {
+    canSelectMany: boolean;
+    openLabel: string;
+    canSelectFolders: boolean;
+    canSelectFiles: boolean;
+}
+
+export class MultiStepInput {
+    async showOpenDialog({ canSelectMany, openLabel, canSelectFolders, canSelectFiles }: OpenDialogParameters): Promise<Uri[] | undefined> {
+        return await new Promise<Uri[] | undefined>((resolve) => {
+            const input: OpenDialogOptions = {
+                canSelectMany,
+                openLabel,
+                canSelectFolders,
+                canSelectFiles
+            };
+
+            void window.showOpenDialog(input).then(folderUris => {
+                if (folderUris && folderUris.length > 0) {
+                    resolve([folderUris[0]]);
+                } else {
+                    resolve(undefined);
+                }
+            });
+        });
+    }
+
+    static async run<T>(start: InputStep, onNavBack?: () => void) {
+        const input = new MultiStepInput();
+        return input.stepThrough(start, onNavBack);
+    }
+
+    private current?: QuickInput;
+    private steps: InputStep[] = [];
+
+    private async stepThrough<T>(start: InputStep, onNavBack?: () => void) {
+        let step: InputStep | void = start;
+        while (step) {
+            this.steps.push(step);
+            if (this.current) {
+                this.current.enabled = false;
+                this.current.busy = true;
+            }
+            try {
+                step = await step(this);
+            } catch (err) {
+                if (err === InputFlowAction.back) {
+                    if (onNavBack) {
+                        onNavBack();  //Currently, step -= 2 passed as onNavBack because of using postfix increment in steps in the input functions
+                    }
+                    this.steps.pop();
+                    step = this.steps.pop();
+                } else if (err === InputFlowAction.resume) {
+                    step = this.steps.pop();
+                } else if (err === InputFlowAction.cancel) {
+                    step = undefined;
+                } else {
+                    throw err;
+                }
+            }
+        }
+        if (this.current) {
+            this.current.dispose();
+        }
+    }
+
+    async showQuickPick<T extends QuickPickItem, P extends QuickPickParameters<T>>({ title, step, totalSteps, items, activeItem, ignoreFocusOut, placeholder, buttons, shouldResume }: P) {
+        const disposables: Disposable[] = [];
+        try {
+            return await new Promise<T | (P extends { buttons: (infer I)[] } ? I : never)>((resolve, reject) => {
+                const input = window.createQuickPick<T>();
+                input.title = title;
+                input.step = step;
+                input.totalSteps = totalSteps;
+                input.ignoreFocusOut = ignoreFocusOut ?? false;
+                input.placeholder = placeholder;
+                input.items = items;
+                input.buttons = [
+                    ...(this.steps.length > 1 ? [QuickInputButtons.Back] : []),
+                    ...(buttons || [])
+                ];
+                disposables.push(
+                    input.onDidTriggerButton(item => {
+                        if (item === QuickInputButtons.Back) {
+                            reject(InputFlowAction.back);
+                        } else {
+                            resolve(<any>item);
+                        }
+                    }),
+                    input.onDidChangeSelection(items => resolve(items[0])),
+                    input.onDidHide(() => {
+                        (async () => {
+                            reject(shouldResume && await shouldResume() ? InputFlowAction.resume : InputFlowAction.cancel);
+                        })()
+                            .catch(reject);
+                    })
+                );
+                if (this.current) {
+                    this.current.dispose();
+                }
+                this.current = input;
+                this.current.show();
+                if (activeItem) {
+                    input.activeItems = [activeItem];
+                }
+            });
+        } finally {
+            disposables.forEach(d => d.dispose());
+        }
+    }
+
+    async showInputBox<P extends InputBoxParameters>({ title, step, totalSteps, value, prompt, validate, buttons, ignoreFocusOut, placeholder, shouldResume }: P) {
+        const disposables: Disposable[] = [];
+        try {
+            return await new Promise<string | (P extends { buttons: (infer I)[] } ? I : never)>((resolve, reject) => {
+                const input = window.createInputBox();
+                input.title = title;
+                input.step = step;
+                input.totalSteps = totalSteps;
+                input.value = value || '';
+                input.prompt = prompt;
+                input.ignoreFocusOut = ignoreFocusOut ?? false;
+                input.placeholder = placeholder;
+                input.buttons = [
+                    ...(this.steps.length > 1 ? [QuickInputButtons.Back] : []),
+                    ...(buttons || [])
+                ];
+                let validating = validate ? validate('') : Promise.resolve(undefined);
+                disposables.push(
+                    input.onDidTriggerButton(item => {
+                        if (item === QuickInputButtons.Back) {
+                            reject(InputFlowAction.back);
+                        } else {
+                            resolve(<any>item);
+                        }
+                    }),
+                    input.onDidAccept(async () => {
+                        const value = input.value;
+                        input.enabled = false;
+                        input.busy = true;
+                        if (!(validate && await validate(value))) {
+                            resolve(value);
+                        }
+                        input.enabled = true;
+                        input.busy = false;
+                    }),
+                    input.onDidChangeValue(async text => {
+                        if (validate) {
+                            const current = validate(text);
+                            validating = current;
+                            const validationMessage = await current;
+                            if (current === validating) {
+                                input.validationMessage = validationMessage;
+                            }
+                        }
+                    }),
+                    input.onDidHide(() => {
+                        (async () => {
+                            reject(shouldResume && await shouldResume() ? InputFlowAction.resume : InputFlowAction.cancel);
+                        })()
+                            .catch(reject);
+                    })
+                );
+                if (this.current) {
+                    this.current.dispose();
+                }
+                this.current = input;
+                this.current.show();
+            });
+        } finally {
+            disposables.forEach(d => d.dispose());
+        }
+    }
+}
